Reorganize files. Changed filepath for ItemStatus.java and PantryItem.java to demo1 folder
